// frontend/src/shared/api/schema.ts
export interface paths {
    "/api/v1/employees/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get Employee */
        get: operations["org_departments_api_get_employee"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/employees/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List Employees */
        get: operations["org_departments_api_list_employees"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/departments/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List Departments */
        get: operations["org_departments_api_list_departments"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/departments/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get Department */
        get: operations["org_departments_api_get_department"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/departments/{department_id}/employees": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List Departments Employees */
        get: operations["org_departments_api_list_departments_employees"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /** EmployeeSchema */
        EmployeeSchema: {
            /** Id */
            id: number;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /** Middle Name */
            middle_name: string | null;
            /** Amount */
            amount: number;
            /**
             * Hire Date
             * Format: date-time
             */
            hire_date: string;
            /** Position Title */
            position_title: string | null;
            /** Department Id */
            department_id: number | null;
        };
        /** EmployeeNotFoundSchema */
        EmployeeNotFoundSchema: {
            /**
             * Ok
             * @constant
             */
            ok: false;
            /**
             * Error Code
             * @constant
             */
            error_code: "employee_not_found";
        };
        /** Input */
        Input: {
            /**
             * Limit
             * @default 100
             */
            limit: number;
            /**
             * Offset
             * @default 0
             */
            offset: number;
        };
        /** PagedEmployeeSchema */
        PagedEmployeeSchema: {
            /** Items */
            items: components["schemas"]["EmployeeSchema"][];
            /** Count */
            count: number;
        };
        /** DepartmentSchema */
        DepartmentSchema: {
            /** Id */
            id: number;
            /** Title */
            title: string;
            /** Parent Id */
            parent_id: number | null;
            /**
             * Children
             * @default []
             */
            children: components["schemas"]["DepartmentSchema"][];
        };
        /** PagedDepartmentSchema */
        PagedDepartmentSchema: {
            /** Items */
            items: components["schemas"]["DepartmentSchema"][];
            /** Count */
            count: number;
        };
        /** DepartmentFoundSchema */
        DepartmentFoundSchema: {
            /**
             * Ok
             * @constant
             */
            ok: false;
            /**
             * Error Code
             * @constant
             */
            error_code: "department_not_found";
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    org_departments_api_get_employee: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["EmployeeSchema"];
                };
            };
            /** @description Gone */
            410: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["EmployeeNotFoundSchema"];
                };
            };
        };
    };
    org_departments_api_list_employees: {
        parameters: {
            query?: {
                limit?: number;
                offset?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["PagedEmployeeSchema"];
                };
            };
        };
    };
    org_departments_api_list_departments: {
        parameters: {
            query?: {
                limit?: number;
                offset?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["PagedDepartmentSchema"];
                };
            };
        };
    };
    org_departments_api_get_department: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["DepartmentSchema"];
                };
            };
            /** @description Gone */
            410: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["DepartmentFoundSchema"];
                };
            };
        };
    };
    org_departments_api_list_departments_employees: {
        parameters: {
            query?: {
                limit?: number;
                offset?: number;
            };
            header?: never;
            path: {
                department_id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["PagedEmployeeSchema"];
                };
            };
        };
    };
}